requirements.txt
pip
numpy
pandas
scipy
mkl
matplotlib
h5py
pytorch
pytest
skorch
joblib
memory_profiler
tqdm
scikit-learn
yaml
pyyaml
git+https://github.com/ostormer/mne-python.git@main#egg=mne
git+https://github.com/ostormer/braindecode.git@master#egg=braindecode
torchplot
